feat: 添加导师系统、人才墙和团队项目功能
主要功能：
- 导师系统：导师申请、导师关系管理、导师评分
- 人才墙：用户可用性管理、人才展示
- 团队项目：团队与项目关联管理
- 标签系统：用户标签、管理员标签管理
- 生态系统：比赛中心、人才广场
- 数据库迁移：添加相关表结构

技术改进：
- 优化前端性能：懒加载、路由预加载
- 改进用户体验：字段规范化、表单验证
- 增强安全性：权限控制、数据验证
- 代码质量：类型定义、错误处理

+++ backend/teamup-server/src/main/java/com/teamup/server/modules/mentor/vo/MentorInfoVO.java
@@ -2,68 +2,21 @@ package com.teamup.server.modules.mentor.vo;
 
 import lombok.Data;
 
-import java.math.BigDecimal;
-
 /**
  * 导师信息VO
  */
 @Data
 public class MentorInfoVO {
-    
     private Long id;
-    
-    /**
-     * 用户名
-     */
     private String username;
-    
-    /**
-     * 真实姓名
-     */
     private String realName;
-    
-    /**
-     * 院系
-     */
     private String department;
-    
-    /**
-     * 专业
-     */
     private String major;
-    
-    /**
-     * 总学员数
-     */
     private Integer totalMentees;
-    
-    /**
-     * 当前活跃学员数
-     */
     private Integer activeMentees;
-    
-    /**
-     * 已完成学员数
-     */
     private Integer completedMentees;
-    
-    /**
-     * 成功培养学员数
-     */
     private Integer successfulMentees;
-    
-    /**
-     * 学员平均信誉分
-     */
-    private BigDecimal averageMenteeScore;
-    
-    /**
-     * 累计奖励积分
-     */
+    private Double averageMenteeScore;
     private Integer totalRewardPoints;
-    
-    /**
-     * 导师评分
-     */
-    private BigDecimal rating;
+    private Double rating;
 }
